Move wrapRegExp to an individual file
